Added Neural Net Model/RData
Added Neural Net Model & RData.

RData is quite large (500MB) and is based on ALL features (SIFT+HOG).
As the file is so large, it will be better to regenerate the
model going forward so as to not bloat the directory. Heuristically
it’s found that we get the lowest test error using 5 neurons in the
hidden layer when using just SIFT features and 3 when using SIFT+HOG
features.